project/requirements.txt
asgiref==3.2.3
Django==3.0.2
django-crispy-forms==1.8.1
mysqlclient==1.4.6
Pillow==7.0.0
pytz==2019.3
six==1.12.0
sqlparse==0.3.0
